+++ docs/.vitepress/config.ts
@@ -24,7 +24,7 @@ export default defineConfig({
       { text: 'Security', link: '/security/overview' },
       { text: 'Reference', link: '/reference/env-vars' },
       {
-        text: 'v0.4.0',
+        text: 'v0.5.0',
         items: [
           { text: 'Changelog', link: '/reference/changelog' },
           { text: 'GitHub', link: 'https://github.com/chr0nzz/gatekeeper' },
@@ -61,6 +61,7 @@ export default defineConfig({
         items: [
           { text: 'Overview', link: '/oidc/provider' },
           { text: 'Managing clients', link: '/admin/managing-clients' },
+          { text: 'Custom claims', link: '/admin/custom-claims' },
         ],
       },
       {